tx-generator: migrate ADTs (WIP3)
