+++ bench/tx-generator/src/Cardano/TxGenerator/Utils.hs
@@ -7,6 +7,8 @@ module  Cardano.TxGenerator.Utils
 
 import           Cardano.Api as Api
 
+import           Cardano.TxGenerator.Types
+
 
 liftAnyEra :: ( forall era. IsCardanoEra era => f1 era -> f2 era ) -> InAnyCardanoEra f1 -> InAnyCardanoEra f2
 liftAnyEra f x = case x of
@@ -17,10 +19,24 @@ liftAnyEra f x = case x of
   InAnyCardanoEra AlonzoEra a  ->   InAnyCardanoEra AlonzoEra $ f a
   InAnyCardanoEra BabbageEra a  ->  InAnyCardanoEra BabbageEra $ f a
 
--- Todo: check sufficient funds and minimumValuePerUtxo
+keyAddress :: forall era. IsShelleyBasedEra era => NetworkId -> SigningKey PaymentKey -> AddressInEra era
+keyAddress networkId k
+  = makeShelleyAddressInEra
+      networkId
+      (PaymentCredentialByKey $ verificationKeyHash $ getVerificationKey k)
+      NoStakeAddress
+
+-- TODO: check sufficient funds and minimumValuePerUtxo
 inputsToOutputsWithFee :: Lovelace -> Int -> [Lovelace] -> [Lovelace]
 inputsToOutputsWithFee fee count inputs = map (quantityToLovelace . Quantity) outputs
   where
     (Quantity totalAvailable) = lovelaceToQuantity $ sum inputs - fee
     (out, rest) = divMod totalAvailable (fromIntegral count)
     outputs = (out + rest) : replicate (count-1) out
+
+includeChange :: Lovelace -> [Lovelace] -> [Lovelace] -> PayWithChange
+includeChange fee spend have = case compare changeValue 0 of
+  GT -> PayWithChange changeValue spend
+  EQ -> PayExact spend
+  LT -> error "genTX: Bad transaction: insufficient funds"
+  where changeValue = sum have - sum spend - fee
